// frontend/src/app/generated-api/model/requestDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RequestDto { 
    /**
     * Der neu zu berechnende Geldbetrag (muss >= 0.00 sein)
     */
    newAmount: number;
    /**
     * Der alte Vergleichsbetrag (optional, muss >= 0.00 sein)
     */
    oldAmount?: number;
}