// frontend/src/app/generated-api/model/responseDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ResponseDto { 
    /**
     * neuer Betrag
     */
    newAmount?: number;
    /**
     * alter Betrag
     */
    oldAmount?: number;
    /**
     * Die optimale Stückelung des neuen Betrags
     */
    newDenomination?: { [key: string]: number; };
    /**
     * Die Differenz der Stückelung im Vergleich zum alten Betrag
     */
    difference?: { [key: string]: number; };
}